refactor : MemberNotFoundException 패키지 이동
