+++ api/src/main/java/com/dnd/sbooky/api/book/UpdateBookUseCase.java
@@ -1,7 +1,7 @@
 package com.dnd.sbooky.api.book;
 
 import com.dnd.sbooky.api.book.exception.BookForbiddenException;
-import com.dnd.sbooky.api.book.exception.MemberNotFoundException;
+import com.dnd.sbooky.api.member.exception.MemberNotFoundException;
 import com.dnd.sbooky.api.book.request.UpdateBookRequest;
 import com.dnd.sbooky.api.support.error.ErrorType;
 import com.dnd.sbooky.core.book.BookEntity;
